requirements.txt
asgiref==3.6.0
certifi==2023.5.7
cloudinary==1.33.0
Django==4.2.1
django-cors-headers==4.0.0
django-filter==23.2
djangorestframework==3.14.0
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
urllib3==1.26.15
